+++ tiger-zion/src/main/java/de/gematik/test/tiger/zion/config/TigerMockResponse.java
@@ -3,13 +3,12 @@ package de.gematik.test.tiger.zion.config;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.*;
 
 @Data
 @AllArgsConstructor(onConstructor_ = @JsonIgnore)
@@ -19,11 +18,18 @@ import lombok.NoArgsConstructor;
 public class TigerMockResponse {
 
     @TigerSkipEvaluation
-    private List<String> requestCriterions;
+    @Builder.Default
+    private List<String> requestCriterions = new ArrayList<>();
+    private ZionRequestMatchDefinition request;
     private TigerMockResponseDescription response;
     private Map<String, TigerMockResponse> nestedResponses = new HashMap<>();
     private Map<String, ZionBackendRequestDescription> backendRequests = new HashMap<>();
     @TigerSkipEvaluation
-    private Map<String, String> assignments;
+    @Builder.Default
+    private Map<String, String> assignments = new HashMap<>();
     private int importance = 0;
+
+    public Optional<ZionRequestMatchDefinition> getRequestOptional() {
+        return Optional.ofNullable(request);
+    }
 }